requirements.txt
PyMuPDF>=1.24.0
python-docx>=1.1.0
requests>=2.31.0
